refactor: move Network (and deps) into separate module
