+++ src/section/mod.rs
@@ -8,9 +8,6 @@
 
 mod elders_info;
 mod member_info;
-mod network;
-mod network_stats;
-mod prefix_map;
 #[allow(clippy::module_inception)]
 mod section;
 mod section_keys;
@@ -24,13 +21,11 @@ pub(crate) use self::elders_info::gen_elders_info;
 pub use self::{
     elders_info::{majority_count, EldersInfo},
     member_info::{MemberInfo, PeerState, MIN_AGE},
-    network_stats::NetworkStats,
     section_keys::{SectionKeyShare, SectionKeysProvider},
     section_peers::SectionPeers,
     section_proof_chain::{ExtendError, SectionProofChain, TrustStatus},
 };
 pub(crate) use self::{
-    network::Network,
     section::Section,
     section_update_barrier::SectionUpdateBarrier,
     shared_state::{SharedState, UpdateSectionKnowledgeAction},
